Fix type in member check and improve SSE emit handling
Changed SQL query result type from i64 to i32 in claim_pending_device_addition to match PostgreSQL's SELECT 1. Updated SseState::emit to treat absence of subscribers as non-fatal, returning Ok even if no receivers are present.

+++ server/src/handlers/claim_pending_device_addition.rs
@@ -135,7 +135,8 @@ pub async fn claim_pending_device_addition(
     }
 
     // Verify caller is a member of the conversation
-    let is_member: Option<(i64,)> = sqlx::query_as(
+    // Note: SELECT 1 returns PostgreSQL INTEGER (4 bytes), so we use i32 to match
+    let is_member: Option<(i32,)> = sqlx::query_as(
         r#"
         SELECT 1
         FROM members
